requirements.txt
django
mock
ujson
pytils
django-annoying
django-eml-email-backend
django-modeldict
textile
django-constance
django-constance-trbs
django-picklefield
django-robots
django-admin-honeypot
django-email-obfuscator
django.js
south
factory_boy
django-factory-boy
webtest
django-webtest
coverage >= 3.0
django-coverage
selenium
redis
django-email-html
cleanweb
django-cleanup
RandomWords==0.1.5
psycopg2
fabric
djangorestframework==2.3.6